pyproject.toml
[tool.poetry]
name = "clients-data-analysis"
version = "0.1.0"
description = ""
authors = ["matheusinit <matheus.oliveira.s@protonmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
openpyxl = "^3.1.5"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
